azure-functions/requirements.txt
azure-functions
supabase>=2.0.0
python-dateutil>=2.8.0
pyyaml>=6.0
